AvlTreeWithoutIndex
Implemented an AVL tree to simulate an editor that does not maintain an
index. On the contrary it maintains a count of number of leaf nodes that
a parent node has. The position of the leaf node is the position of the
line number. I have mixed the concepts of Height balance and Weight
balance tree to achieve this.